+++ source/renderer/app/containers/status/DaedalusDiagnosticsDialog.js
@@ -50,7 +50,8 @@ export default class DaedalusDiagnosticsDialog extends Component<Props> {
       environment,
       diskSpaceAvailable,
       tlsConfig,
-      cardanoNodeID,
+      cardanoNodePID,
+      cardanoWalletPID,
       stateDirectoryPath,
       getNetworkClockRequest,
     } = networkStatus;
@@ -71,6 +72,7 @@ export default class DaedalusDiagnosticsDialog extends Component<Props> {
       mainProcessID,
       isBlankScreenFixActive,
       nodeVersion,
+      apiVersion,
     } = environment;
 
     const coreInfo = {
@@ -79,9 +81,11 @@ export default class DaedalusDiagnosticsDialog extends Component<Props> {
       daedalusMainProcessID: mainProcessID,
       daedalusStateDirectoryPath: stateDirectoryPath,
       isBlankScreenFixActive,
-      cardanoVersion: nodeVersion,
-      cardanoProcessID: cardanoNodeID,
-      cardanoAPIPort: tlsConfig ? tlsConfig.port : 0,
+      cardanoNodeVersion: nodeVersion,
+      cardanoNodePID,
+      cardanoWalletVersion: apiVersion,
+      cardanoWalletPID,
+      cardanoWalletApiPort: tlsConfig ? tlsConfig.port : 0,
       cardanoNetwork: network,
       cardanoRawNetwork: rawNetwork,
     };